% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Affy1_Distance_Final}
\alias{Affy1_Distance_Final}
\title{Targets for the microRNA analyzed from Affy1 plaform using Distance.}
\format{A data frame with 53399 rows and 8 variables:
\describe{
  \item{miRNA}{miRNA name, miRNA symbol}
  \item{GeneSymbol}{Gene name, in Gene Symbol}
  \item{FBGN}{Gene name, in FlybaseID}
  \item{CGID}{Gene name, in CGID}
  \item{Score}{Computed Score, in float}
  \item{GeneFunction}{Gene Functions, from Flybase}
  \item{experiments}{Experiments, from ArrayExpress}
  \item{TargetDatabases}{Target Database Name, from TargetDatabases}
}}
\usage{
Affy1_Distance_Final
}
\description{
A precomputed dataset containing the targets, scores and other 
attributes of 83 intragenic microRNAs using Distance Correlation 
for plaform Affymetrix 1.
}
\keyword{datasets}

