% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_miRNA_HostGene.r
\name{extract_HostGene}
\alias{extract_HostGene}
\title{Extract Host Gene for a given Intragenic miRNA.}
\usage{
extract_HostGene(miRNA)
}
\arguments{
\item{miRNA}{A String containing the miRNA name.}
}
\value{
genf, a character string or vector containing
        Host gene for the Intragenic miRNA.
}
\description{
Extract Host Gene for a given Intragenic miRNA.
}
\examples{
miRNA="dme-miR-12"
extract_HostGene(miRNA)
}

