% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_miRNA_HostGene.r
\name{extract_intragenic_miR}
\alias{extract_intragenic_miR}
\title{Extract Intragenic miRNA for a given Host gene.}
\usage{
extract_intragenic_miR(gene)
}
\arguments{
\item{gene}{character. Gene Symbol.}
}
\value{
miRf, a character string or vector containing
        Intragenic miRNA for the Host Gene.
}
\description{
Extract Intragenic miRNA for a given Host gene.
}
\examples{
gene="Gmap"
extract_intragenic_miR(gene)
}

