% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntraMirExploreR.r
\docType{package}
\name{IntramiRExploreR}
\alias{IntramiRExploreR}
\alias{IntramiRExploreR-package}
\title{IntramiRExploreR:Prediction of targets for Intragenic 
miRNA in Drosophila.}
\description{
Prediction of targets for Drosophila Intragenic microRNAs
and Integrated approach using Gene Ontology and Networking tools.
}
\examples{
\dontrun{
gene='Gmap'
extract_intragenic_miR(gene)
}
}

