% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{miRNA_summary_DB}
\alias{miRNA_summary_DB}
\title{Contains the summary for the intragenic miRNA.}
\format{A data frame with 257 rows and 6 variables:
\describe{
  \item{miRNA}{miRNA name, miRNA symbol}
  \item{Intragenic}{Responsee, in boolean}
  \item{Intergenic}{Responsee, in boolean}
  \item{Gene}{miRNA name, miRNA symbol}
  \item{Type.of.HostGene.mRNA.lncRNA.}{Type of Hostgene}
  \item{Notes}{Comments about the miRNA}
}}
\usage{
miRNA_summary_DB
}
\description{
A dataset containing the summary for the intragenic miRNA.
}
\keyword{datasets}

